Add servercontroller class.
